// src/app/core/api/model/adminPatchPaymentRequestDto.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminPatchPaymentRequestDto { 
    status: AdminPatchPaymentRequestDto.StatusEnum;
    amountCents?: number;
    paymentMode?: AdminPatchPaymentRequestDto.PaymentModeEnum;
    checkNumber?: string;
    comment?: string;
}
export namespace AdminPatchPaymentRequestDto {
    export const StatusEnum = {
        DepositPending: 'DEPOSIT_PENDING',
        DepositPaid: 'DEPOSIT_PAID',
        Exempt: 'EXEMPT',
        Waived: 'WAIVED',
        Refunded: 'REFUNDED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const PaymentModeEnum = {
        Cash: 'CASH',
        Check: 'CHECK'
    } as const;
    export type PaymentModeEnum = typeof PaymentModeEnum[keyof typeof PaymentModeEnum];
}