// src/app/core/api/model/adminPatchReservationStatusRequestDto.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminPatchReservationStatusRequestDto { 
    status: AdminPatchReservationStatusRequestDto.StatusEnum;
    comment?: string;
}
export namespace AdminPatchReservationStatusRequestDto {
    export const StatusEnum = {
        PendingValidation: 'PENDING_VALIDATION',
        Confirmed: 'CONFIRMED',
        Cancelled: 'CANCELLED',
        Rejected: 'REJECTED',
        PendingDocument: 'PENDING_DOCUMENT'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}