// src/app/core/api/model/reservationSummaryResponseDto.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocalTime } from './localTime';


export interface ReservationSummaryResponseDto { 
    id?: string;
    resourceName?: string;
    resourceType?: ReservationSummaryResponseDto.ResourceTypeEnum;
    date?: string;
    slotStart?: LocalTime;
    slotEnd?: LocalTime;
    status?: ReservationSummaryResponseDto.StatusEnum;
    depositStatus?: ReservationSummaryResponseDto.DepositStatusEnum;
    depositAmountCents?: number;
    depositAmountFullCents?: number;
    discountLabel?: string;
    createdAt?: string;
    userName?: string;
}
export namespace ReservationSummaryResponseDto {
    export const ResourceTypeEnum = {
        Immobilier: 'IMMOBILIER',
        Mobilier: 'MOBILIER'
    } as const;
    export type ResourceTypeEnum = typeof ResourceTypeEnum[keyof typeof ResourceTypeEnum];
    export const StatusEnum = {
        PendingValidation: 'PENDING_VALIDATION',
        Confirmed: 'CONFIRMED',
        Cancelled: 'CANCELLED',
        Rejected: 'REJECTED',
        PendingDocument: 'PENDING_DOCUMENT'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const DepositStatusEnum = {
        DepositPending: 'DEPOSIT_PENDING',
        DepositPaid: 'DEPOSIT_PAID',
        Exempt: 'EXEMPT',
        Waived: 'WAIVED',
        Refunded: 'REFUNDED'
    } as const;
    export type DepositStatusEnum = typeof DepositStatusEnum[keyof typeof DepositStatusEnum];
}