// src/app/core/api/model/adminPaymentHistoryEntryResponseDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminPaymentHistoryEntryResponseDto {
  status?: AdminPaymentHistoryEntryResponseDto.StatusEnum;
  amountCents?: number;
  paymentMode?: AdminPaymentHistoryEntryResponseDto.PaymentModeEnum;
  checkNumber?: string;
  comment?: string;
  updatedByName?: string;
  updatedAt?: string;
}
export namespace AdminPaymentHistoryEntryResponseDto {
  export const StatusEnum = {
    DepositPending: 'DEPOSIT_PENDING',
    DepositPaid: 'DEPOSIT_PAID',
    Exempt: 'EXEMPT',
    Waived: 'WAIVED',
    Refunded: 'REFUNDED',
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
  export const PaymentModeEnum = {
    Cash: 'CASH',
    Check: 'CHECK',
  } as const;
  export type PaymentModeEnum = (typeof PaymentModeEnum)[keyof typeof PaymentModeEnum];
}