// src/app/core/api/model/pagedResponseReservationSummaryResponseDto.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReservationSummaryResponseDto } from './reservationSummaryResponseDto';


export interface PagedResponseReservationSummaryResponseDto { 
    content?: Array<ReservationSummaryResponseDto>;
    totalElements?: number;
    totalPages?: number;
    page?: number;
    size?: number;
}